Move deserialization tests
