+++ CMakeLists.txt
@@ -306,13 +306,13 @@ if(CATKIN_ENABLE_TESTING)
     ${catkin_LIBRARIES}
   )
 
-  catkin_add_gtest(unittest_monitoring_frame_serialization
+  catkin_add_gtest(unittest_monitoring_frame_serialization_deserialization
     src/monitoring_frame_msg.cpp
     src/monitoring_frame_deserialization.cpp
-    test/unit_tests/unittest_monitoring_frame_serialization.cpp
+    test/unit_tests/unittest_monitoring_frame_serialization_deserialization.cpp
     test/src/monitoring_frame_serialization.cpp
   )
-  target_link_libraries(unittest_monitoring_frame_serialization
+  target_link_libraries(unittest_monitoring_frame_serialization_deserialization
     ${catkin_LIBRARIES}
   )
